DistQ now uses redis for keeping state entirelly.
The queue of jobs is represented in redis as a job_pool hash, where
each key is a unique key of a job and each value is the serialized
version of the job.

Next that there are sorted sets, one for each "state" a job can be
in (as of now, "schedule", "ready", "working", "done").

When a job first comes in it's added to the queue in the "schedule"
state, where the score of the schedule sorted set is the seconds
since the epoch of the time it should be run.

As a job moves through various states it's removed from one sorted
set and into another, all atomically so we never lose the full picture.