Bake job states into the redis scripts. [#7] [#10]
Before, there was a generic qsetstate script for setting arbitrary
states.  Now, there are just a handful of lua scripts that are well
aware of which states it will involve.

The states are represented as simple sets, with the exception of the
initial "scheduled" state, which is still a sorted set where "score" is
the epoch seconds scheduled time of the job.