Update Arbiter to keep track of its capacity.
This way we don't end up filling up the 'ready' IPC queue with
absolutely every scheduled job if there's a backup.  Now what will
happen is that the jobs that can't be grabbed yet will stay in the
'scheduled' state.  The current epoch seconds minus the lowest score in
the sorted set represent the queue's delay.