Support multilple named queues.
The @job decorator is updated to require naming a queue that the job
belongs in, there is no longer a default queue named "queue".

The bulk of the update is to the qpop.lua script.  Since a consumer asks
for a set number of jobs with the smallest scores across scheduled sets
the complication of the operation goes up a bit.